Feat : 프로젝트 구조 수정
프로젝트 구조를 계층형에서 도메인형으로 수정
팀 프로젝트에 더 적합한 구조로 수정

Resolves : #14
