+++ src/main/java/HealthMe/HealthMe/domain/userbodyinformation/domain/UserBodyInformation.java
@@ -1,8 +1,10 @@
-package HealthMe.HealthMe.model;
+package HealthMe.HealthMe.domain.userbodyinformation.domain;
 
-
+import HealthMe.HealthMe.domain.user.domain.User;
 import jakarta.persistence.*;
+import lombok.AllArgsConstructor;
 import lombok.Getter;
+import lombok.NoArgsConstructor;
 import lombok.Setter;
 
 /**
@@ -23,27 +25,22 @@ import lombok.Setter;
  */
 @Getter
 @Setter
-@Entity(name = "PRESET")
-public class Preset {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "PRESET_ID")
-    private Long id;    // pk를 위한 id
+@NoArgsConstructor
+@AllArgsConstructor
+@Entity(name = "USER_BODY_INFORMATION")
+public class UserBodyInformation {
+    @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name="USER_BODY_INFORMATION_ID")
+    private Long id;        // pk를 위한 id
 
-    private String email;   // user 구분을 위한 email
+    @Column(unique = true)
+    private String email;   // user 구분을 위한 이메일
+    private String height;  // 키
+    private String weight;  // 몸무게
+    private String gender;  // 성별
+    private Integer age;    // 나이
 
-    private Long presetNumber;  // 프리셋 번호
-
-    private String exerciseName;    // 운동 명
-    private Double weight;  // 무게
-
-    private Integer setCount;   // 세트 수
-
-    private Integer repetitionCount;    // 반복 횟수
-
-    private Integer restTime;   // 쉬는 시간 (나중에 시간 자료형으로 변경 계획 있음)
-
-    @ManyToOne
-    @JoinColumn(name = "USER_ID")
+    @OneToOne
+    @JoinColumn(name="USER_ID")
     private User user;
 }
